requirements.txt
python-telegram-bot==13.8.1
Pillow
toml
